data-pipeline/requirements.txt
# Core Data Processing
pandas>=2.1.0,<2.3.0
pyarrow>=12.0.0,<14.0.0
numpy>=1.21.0,<2.0.0

# AWS Integration
boto3>=1.28.35,<2.0.0

# Orchestration
prefect==3.5.0

# Configuration Management
python-dotenv>=1.0.0,<2.0.0